@@ -5,13 +5,13 @@ version: 1.0.0
 homepage: https://github.com/fayaz07/ots
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  connectivity_plus: ^1.0.6
+  connectivity_plus: ^6.0.5
 
 dev_dependencies:
   flutter_test:
